Trait pallet_housing_fund::storage::types::EncodeLikeTuple

pub trait EncodeLikeTuple<T>: Sealed { }

Marker trait to indicate that each element in the tuple encodes like the corresponding element in another tuple.

This trait is sealed.