Trait pallet_housing_fund::storage::types::TupleToEncodedIter

pub trait TupleToEncodedIter {
    fn to_encoded_iter(&self) -> IntoIter<Vec<u8, Global>, Global>;
}

Trait to indicate that a tuple can be converted into an iterator of a vector of encoded bytes.

Required Methods

fn to_encoded_iter(&self) -> IntoIter<Vec<u8, Global>, Global>