Struct pallet_housing_fund::CountedStorageMap

pub struct CountedStorageMap<Prefix, Hasher, Key, Value, QueryKind = OptionQuery, OnEmpty = GetDefault, MaxValues = GetDefault>(_);

A wrapper around a StorageMap and a StorageValue<Value=u32> to keep track of how many items are in a map, without needing to iterate all the values.

This storage item has additional storage read and write overhead when manipulating values compared to a regular storage map.

For functions where we only add or remove a value, a single storage read is needed to check if that value already exists. For mutate functions, two storage reads are used to check if the value existed before and after the mutation.

Whenever the counter needs to be updated, an additional read and write occurs to update that counter.

Implementations

impl<Prefix, Hasher, Key, Value, QueryKind, OnEmpty, MaxValues> CountedStorageMap<Prefix, Hasher, Key, Value, QueryKind, OnEmpty, MaxValues>where
Prefix: CountedStorageMapInstance,
Hasher: StorageHasher,
Key: FullCodec,
Value: FullCodec,
QueryKind: QueryKindTrait<Value, OnEmpty>,
OnEmpty: 'static + Get<<QueryKind as QueryKindTrait<Value, OnEmpty>>::Query>,
MaxValues: Get<Option<u32>>,

pub fn counter_storage_final_key() -> [u8; 32]

The key used to store the counter of the map.

pub fn map_storage_final_prefix() -> Vec<u8, Global>

The prefix used to generate the key of the map.

pub fn hashed_key_for<KeyArg>(key: KeyArg) -> Vec<u8, Global>where
KeyArg: EncodeLike<Key>,

Get the storage key used to fetch a value corresponding to a specific key.

pub fn contains_key<KeyArg>(key: KeyArg) -> boolwhere
KeyArg: EncodeLike<Key>,

Does the value (explicitly) exist in storage?

pub fn get<KeyArg>(
key: KeyArg
) -> <QueryKind as QueryKindTrait<Value, OnEmpty>>::Querywhere
KeyArg: EncodeLike<Key>,

Load the value associated with the given key from the map.

pub fn try_get<KeyArg>(key: KeyArg) -> Result<Value, ()>where
KeyArg: EncodeLike<Key>,

Try to get the value for the given key from the map.

Returns Ok if it exists, Err if not.

pub fn set<KeyArg>(
key: KeyArg,
q: <QueryKind as QueryKindTrait<Value, OnEmpty>>::Query
)where
KeyArg: EncodeLike<Key>,

Store or remove the value to be associated with key so that get returns the query.

pub fn swap<KeyArg1, KeyArg2>(key1: KeyArg1, key2: KeyArg2)where
KeyArg1: EncodeLike<Key>,
KeyArg2: EncodeLike<Key>,

Swap the values of two keys.

pub fn insert<KeyArg, ValArg>(key: KeyArg, val: ValArg)where
KeyArg: EncodeLike<Key> + Clone,
ValArg: EncodeLike<Value>,

Store a value to be associated with the given key from the map.

pub fn remove<KeyArg>(key: KeyArg)where
KeyArg: EncodeLike<Key> + Clone,

Remove the value under a key.

pub fn mutate<KeyArg, R, F>(key: KeyArg, f: F) -> Rwhere
KeyArg: EncodeLike<Key> + Clone,
F: FnOnce(&mut <QueryKind as QueryKindTrait<Value, OnEmpty>>::Query) -> R,

Mutate the value under a key.

pub fn try_mutate<KeyArg, R, E, F>(key: KeyArg, f: F) -> Result<R, E>where
KeyArg: EncodeLike<Key> + Clone,
F: FnOnce(&mut <QueryKind as QueryKindTrait<Value, OnEmpty>>::Query) -> Result<R, E>,

Mutate the item, only if an Ok value is returned.

pub fn mutate_exists<KeyArg, R, F>(key: KeyArg, f: F) -> Rwhere
KeyArg: EncodeLike<Key> + Clone,
F: FnOnce(&mut Option<Value>) -> R,

Mutate the value under a key. Deletes the item if mutated to a None.

pub fn try_mutate_exists<KeyArg, R, E, F>(key: KeyArg, f: F) -> Result<R, E>where
KeyArg: EncodeLike<Key> + Clone,
F: FnOnce(&mut Option<Value>) -> Result<R, E>,

Mutate the item, only if an Ok value is returned. Deletes the item if mutated to a None. f will always be called with an option representing if the storage item exists (Some<V>) or if the storage item does not exist (None), independent of the QueryType.

pub fn take<KeyArg>(
key: KeyArg
) -> <QueryKind as QueryKindTrait<Value, OnEmpty>>::Querywhere
KeyArg: EncodeLike<Key> + Clone,

Take the value under a key.

pub fn append<Item, EncodeLikeItem, EncodeLikeKey>(
key: EncodeLikeKey,
item: EncodeLikeItem
)where
EncodeLikeKey: EncodeLike<Key> + Clone,
Item: Encode,
EncodeLikeItem: EncodeLike<Item>,
Value: StorageAppend<Item>,

Append the given items to the value in the storage.

Value is required to implement codec::EncodeAppend.

Warning

If the storage item is not encoded properly, the storage will be overwritten and set to [item]. Any default value set for the storage item will be ignored on overwrite.

pub fn decode_len<KeyArg>(key: KeyArg) -> Option<usize>where
KeyArg: EncodeLike<Key>,
Value: StorageDecodeLength,

Read the length of the storage value without decoding the entire value under the given key.

Value is required to implement StorageDecodeLength.

If the value does not exists or it fails to decode the length, None is returned. Otherwise Some(len) is returned.

Warning

None does not mean that get() does not return a value. The default value is completly ignored by this function.

pub fn migrate_key<OldHasher, KeyArg>(key: KeyArg) -> Option<Value>where
OldHasher: StorageHasher,
KeyArg: EncodeLike<Key>,

Migrate an item with the given key from a defunct OldHasher to the current hasher.

If the key doesn’t exist, then it’s a no-op. If it does, then it returns its value.

pub fn remove_all()

👎Deprecated:
Use clear instead

Remove all values in the map.

pub fn clear(limit: u32, maybe_cursor: Option<&[u8]>) -> MultiRemovalResults

Attempt to remove all items from the map.

Returns MultiRemovalResults to inform about the result. Once the resultant maybe_cursor field is None, then no further items remain to be deleted.

NOTE: After the initial call for any given map, it is important that no further items are inserted into the map. If so, then the map may not be empty when the resultant maybe_cursor is None.

Limit

A limit must always be provided through in order to cap the maximum amount of deletions done in a single call. This is one fewer than the maximum number of backend iterations which may be done by this operation and as such represents the maximum number of backend deletions which may happen. A limit of zero implies that no keys will be deleted, though there may be a single iteration done.

Cursor

A cursor may be passed in to this operation with maybe_cursor. None should only be passed once (in the initial call) for any given storage map. Subsequent calls operating on the same map should always pass Some, and this should be equal to the previous call result’s maybe_cursor field.

pub fn iter_values() -> PrefixIterator<Value, OnRemovalCounterUpdate<Prefix>>

Iter over all value of the storage.

NOTE: If a value failed to decode because storage is corrupted then it is skipped.

pub fn translate_values<OldValue, F>(f: F)where
OldValue: Decode,
F: FnMut(OldValue) -> Option<Value>,

Translate the values of all elements by a function f, in the map in no particular order.

By returning None from f for an element, you’ll remove it from the map.

NOTE: If a value fail to decode because storage is corrupted then it is skipped.

Warning

This function must be used with care, before being updated the storage still contains the old type, thus other calls (such as get) will fail at decoding it.

Usage

This would typically be called inside the module implementation of on_runtime_upgrade.

pub fn try_append<KArg, Item, EncodeLikeItem>(
key: KArg,
item: EncodeLikeItem
) -> Result<(), ()>where
KArg: EncodeLike<Key> + Clone,
Item: Encode,
EncodeLikeItem: EncodeLike<Item>,
Value: StorageTryAppend<Item>,

Try and append the given item to the value in the storage.

Is only available if Value of the storage implements StorageTryAppend.

pub fn initialize_counter() -> u32

Initialize the counter with the actual number of items in the map.

This function iterates through all the items in the map and sets the counter. This operation can be very heavy, so use with caution.

Returns the number of items in the map which is used to set the counter.

pub fn count() -> u32

Return the count.

impl<Prefix, Hasher, Key, Value, QueryKind, OnEmpty, MaxValues> CountedStorageMap<Prefix, Hasher, Key, Value, QueryKind, OnEmpty, MaxValues>where
Prefix: CountedStorageMapInstance,
Hasher: StorageHasher + ReversibleStorageHasher,
Key: FullCodec,
Value: FullCodec,
QueryKind: QueryKindTrait<Value, OnEmpty>,
OnEmpty: 'static + Get<<QueryKind as QueryKindTrait<Value, OnEmpty>>::Query>,
MaxValues: Get<Option<u32>>,

pub fn iter() -> PrefixIterator<(Key, Value), OnRemovalCounterUpdate<Prefix>>

Enumerate all elements in the map in no particular order.

If you alter the map while doing this, you’ll get undefined results.

pub fn drain() -> PrefixIterator<(Key, Value), OnRemovalCounterUpdate<Prefix>>

Remove all elements from the map and iterate through them in no particular order.

If you add elements to the map while doing this, you’ll get undefined results.

pub fn translate<O, F>(f: F)where
O: Decode,
F: FnMut(Key, O) -> Option<Value>,

Translate the values of all elements by a function f, in the map in no particular order.

By returning None from f for an element, you’ll remove it from the map.

NOTE: If a value fail to decode because storage is corrupted then it is skipped.

pub fn iter_from(
starting_raw_key: Vec<u8, Global>
) -> PrefixIterator<(Key, Value), OnRemovalCounterUpdate<Prefix>>

Enumerate all elements in the counted map after a specified starting_raw_key in no particular order.

If you alter the map while doing this, you’ll get undefined results.

pub fn iter_keys() -> KeyPrefixIterator<Key>

Enumerate all keys in the counted map.

If you alter the map while doing this, you’ll get undefined results.