Trait pallet_housing_fund::pallet::Config

pub trait Config: Config + Config {
    type Event: From<Event<Self>> + IsType<<Self as Config>::Event>;
    type LocalCurrency: Currency<Self::AccountId> + ReservableCurrency<Self::AccountId>;
    type MinContribution: Get<BalanceOf<Self>>;
    type FundThreshold: Get<BalanceOf<Self>>;
    type MaxFundContribution: Get<BalanceOf<Self>>;
    type MaxInvestorPerHouse: Get<u32>;
    type PalletId: Get<PalletId>;
    type WeightInfo: WeightInfo;
}

Configure the pallet by specifying the parameters and types on which it depends.

Required Associated Types

type Event: From<Event<Self>> + IsType<<Self as Config>::Event>

Because this pallet emits events, it depends on the runtime’s definition of an event.

type LocalCurrency: Currency<Self::AccountId> + ReservableCurrency<Self::AccountId>

type MinContribution: Get<BalanceOf<Self>>

type FundThreshold: Get<BalanceOf<Self>>

type MaxFundContribution: Get<BalanceOf<Self>>

type MaxInvestorPerHouse: Get<u32>

type PalletId: Get<PalletId>

type WeightInfo: WeightInfo

Weight information for extrinsics in this pallet.