Struct pallet_housing_fund::storage::types::StorageNMap

pub struct StorageNMap<Prefix, Key, Value, QueryKind = OptionQuery, OnEmpty = GetDefault, MaxValues = GetDefault>(_);

A type that allow to store values for an arbitrary number of keys in the form of (Key<Hasher1, key1>, Key<Hasher2, key2>, ..., Key<HasherN, keyN>).

Each value is stored at:

Twox128(Prefix::pallet_prefix())
		++ Twox128(Prefix::STORAGE_PREFIX)
		++ Hasher1(encode(key1))
		++ Hasher2(encode(key2))
	++ ...
	++ HasherN(encode(keyN))

Warning

If the keys are not trusted (e.g. can be set by a user), a cryptographic hasher such as blake2_128_concat must be used for the key hashers. Otherwise, other values in storage can be compromised.

Implementations

impl<Prefix, Key, Value, QueryKind, OnEmpty, MaxValues> StorageNMap<Prefix, Key, Value, QueryKind, OnEmpty, MaxValues>where
Prefix: StorageInstance,
Key: KeyGenerator,
Value: FullCodec,
QueryKind: QueryKindTrait<Value, OnEmpty>,
OnEmpty: 'static + Get<<QueryKind as QueryKindTrait<Value, OnEmpty>>::Query>,
MaxValues: Get<Option<u32>>,

pub fn hashed_key_for<KArg>(key: KArg) -> Vec<u8, Global> where
KArg: EncodeLikeTuple<<Key as KeyGenerator>::KArg> + TupleToEncodedIter,

Get the storage key used to fetch a value corresponding to a specific key.

pub fn contains_key<KArg>(key: KArg) -> boolwhere
KArg: EncodeLikeTuple<<Key as KeyGenerator>::KArg> + TupleToEncodedIter,

Does the value (explicitly) exist in storage?

pub fn get<KArg>(
key: KArg
) -> <QueryKind as QueryKindTrait<Value, OnEmpty>>::Querywhere
KArg: EncodeLikeTuple<<Key as KeyGenerator>::KArg> + TupleToEncodedIter,

Load the value associated with the given key from the map.

pub fn try_get<KArg>(key: KArg) -> Result<Value, ()>where
KArg: EncodeLikeTuple<<Key as KeyGenerator>::KArg> + TupleToEncodedIter,

Try to get the value for the given key from the map.

Returns Ok if it exists, Err if not.

pub fn set<KArg>(
key: KArg,
query: <QueryKind as QueryKindTrait<Value, OnEmpty>>::Query
)where
KArg: EncodeLikeTuple<<Key as KeyGenerator>::KArg> + TupleToEncodedIter,

Store or remove the value to be associated with key so that get returns the query.

pub fn take<KArg>(
key: KArg
) -> <QueryKind as QueryKindTrait<Value, OnEmpty>>::Querywhere
KArg: EncodeLikeTuple<<Key as KeyGenerator>::KArg> + TupleToEncodedIter,

Take a value from storage, removing it afterwards.

pub fn swap<KOther, KArg1, KArg2>(key1: KArg1, key2: KArg2)where
KOther: KeyGenerator,
KArg1: EncodeLikeTuple<<Key as KeyGenerator>::KArg> + TupleToEncodedIter,
KArg2: EncodeLikeTuple<<KOther as KeyGenerator>::KArg> + TupleToEncodedIter,

Swap the values of two key-pairs.

pub fn insert<KArg, VArg>(key: KArg, val: VArg)where
KArg: EncodeLikeTuple<<Key as KeyGenerator>::KArg> + TupleToEncodedIter,
VArg: EncodeLike<Value>,

Store a value to be associated with the given keys from the map.

pub fn remove<KArg>(key: KArg)where
KArg: EncodeLikeTuple<<Key as KeyGenerator>::KArg> + TupleToEncodedIter,

Remove the value under the given keys.

pub fn remove_prefix<KP>(
partial_key: KP,
limit: Option<u32>
) -> KillStorageResultwhere
Key: HasKeyPrefix<KP>,

👎Deprecated: Use clear_prefix instead

Remove all values starting with partial_key in the overlay and up to limit in the backend.

All values in the client overlay will be deleted, if there is some limit then up to limit values are deleted from the client backend, if limit is none then all values in the client backend are deleted.

Note

Calling this multiple times per block with a limit set leads always to the same keys being removed and the same result being returned. This happens because the keys to delete in the overlay are not taken into account when deleting keys in the backend.

pub fn clear_prefix<KP>(
partial_key: KP,
limit: u32,
maybe_cursor: Option<&[u8]>
) -> MultiRemovalResultswhere
Key: HasKeyPrefix<KP>,

Attempt to remove items from the map matching a partial_key prefix.

Returns MultiRemovalResults to inform about the result. Once the resultant maybe_cursor field is None, then no further items remain to be deleted.

NOTE: After the initial call for any given map, it is important that no further items are inserted into the map which match the partial key. If so, then the map may not be empty when the resultant maybe_cursor is None.

Limit

A limit must be provided in order to cap the maximum amount of deletions done in a single call. This is one fewer than the maximum number of backend iterations which may be done by this operation and as such represents the maximum number of backend deletions which may happen. A limit of zero implies that no keys will be deleted, though there may be a single iteration done.

Cursor

A cursor may be passed in to this operation with maybe_cursor. None should only be passed once (in the initial call) for any given storage map and partial_key. Subsequent calls operating on the same map/partial_key should always pass Some, and this should be equal to the previous call result’s maybe_cursor field.

pub fn iter_prefix_values<KP>(partial_key: KP) -> PrefixIterator<Value, ()> where
Key: HasKeyPrefix<KP>,

Iterate over values that share the first key.

pub fn mutate<KArg, R, F>(key: KArg, f: F) -> Rwhere
KArg: EncodeLikeTuple<<Key as KeyGenerator>::KArg> + TupleToEncodedIter,
F: FnOnce(&mut <QueryKind as QueryKindTrait<Value, OnEmpty>>::Query) -> R,

Mutate the value under the given keys.

pub fn try_mutate<KArg, R, E, F>(key: KArg, f: F) -> Result<R, E>where
KArg: EncodeLikeTuple<<Key as KeyGenerator>::KArg> + TupleToEncodedIter,
F: FnOnce(&mut <QueryKind as QueryKindTrait<Value, OnEmpty>>::Query) -> Result<R, E>,

Mutate the value under the given keys when the closure returns Ok.

pub fn mutate_exists<KArg, R, F>(key: KArg, f: F) -> Rwhere
KArg: EncodeLikeTuple<<Key as KeyGenerator>::KArg> + TupleToEncodedIter,
F: FnOnce(&mut Option<Value>) -> R,

Mutate the value under the given keys. Deletes the item if mutated to a None.

pub fn try_mutate_exists<KArg, R, E, F>(key: KArg, f: F) -> Result<R, E>where
KArg: EncodeLikeTuple<<Key as KeyGenerator>::KArg> + TupleToEncodedIter,
F: FnOnce(&mut Option<Value>) -> Result<R, E>,

Mutate the item, only if an Ok value is returned. Deletes the item if mutated to a None. f will always be called with an option representing if the storage item exists (Some<V>) or if the storage item does not exist (None), independent of the QueryType.

pub fn append<Item, EncodeLikeItem, KArg>(key: KArg, item: EncodeLikeItem)where
KArg: EncodeLikeTuple<<Key as KeyGenerator>::KArg> + TupleToEncodedIter,
Item: Encode,
EncodeLikeItem: EncodeLike<Item>,
Value: StorageAppend<Item>,

Append the given item to the value in the storage.

Value is required to implement StorageAppend.

Warning

If the storage item is not encoded properly, the storage will be overwritten and set to [item]. Any default value set for the storage item will be ignored on overwrite.

pub fn decode_len<KArg>(key: KArg) -> Option<usize>where
KArg: EncodeLikeTuple<<Key as KeyGenerator>::KArg> + TupleToEncodedIter,
Value: StorageDecodeLength,

Read the length of the storage value without decoding the entire value under the given key1 and key2.

Value is required to implement StorageDecodeLength.

If the value does not exists or it fails to decode the length, None is returned. Otherwise Some(len) is returned.

Warning

None does not mean that get() does not return a value. The default value is completly ignored by this function.

pub fn migrate_keys<KArg>(
key: KArg,
hash_fns: <Key as KeyGenerator>::HArg
) -> Option<Value>where
KArg: EncodeLikeTuple<<Key as KeyGenerator>::KArg> + TupleToEncodedIter,

Migrate an item with the given key from defunct hash_fns to the current hashers.

If the key doesn’t exist, then it’s a no-op. If it does, then it returns its value.

pub fn remove_all(limit: Option<u32>) -> KillStorageResult

👎Deprecated: Use clear instead

Remove all values in the overlay and up to limit in the backend.

All values in the client overlay will be deleted, if there is some limit then up to limit values are deleted from the client backend, if limit is none then all values in the client backend are deleted.

Note

Calling this multiple times per block with a limit set leads always to the same keys being removed and the same result being returned. This happens because the keys to delete in the overlay are not taken into account when deleting keys in the backend.

pub fn clear(limit: u32, maybe_cursor: Option<&[u8]>) -> MultiRemovalResults

Attempt to remove all items from the map.

Returns MultiRemovalResults to inform about the result. Once the resultant maybe_cursor field is None, then no further items remain to be deleted.

NOTE: After the initial call for any given map, it is important that no further items are inserted into the map. If so, then the map may not be empty when the resultant maybe_cursor is None.

Limit

A limit must always be provided through in order to cap the maximum amount of deletions done in a single call. This is one fewer than the maximum number of backend iterations which may be done by this operation and as such represents the maximum number of backend deletions which may happen. A limit of zero implies that no keys will be deleted, though there may be a single iteration done.

Cursor

A cursor may be passed in to this operation with maybe_cursor. None should only be passed once (in the initial call) for any given storage map. Subsequent calls operating on the same map should always pass Some, and this should be equal to the previous call result’s maybe_cursor field.

pub fn iter_values() -> PrefixIterator<Value, ()>

Iter over all value of the storage.

NOTE: If a value failed to decode because storage is corrupted then it is skipped.

pub fn translate_values<OldValue, F>(f: F)where
OldValue: Decode,
F: FnMut(OldValue) -> Option<Value>,

Translate the values of all elements by a function f, in the map in no particular order. By returning None from f for an element, you’ll remove it from the map.

NOTE: If a value fail to decode because storage is corrupted then it is skipped.

Warning

This function must be used with care, before being updated the storage still contains the old type, thus other calls (such as get) will fail at decoding it.

Usage

This would typically be called inside the module implementation of on_runtime_upgrade.

impl<Prefix, Key, Value, QueryKind, OnEmpty, MaxValues> StorageNMap<Prefix, Key, Value, QueryKind, OnEmpty, MaxValues>where
Prefix: StorageInstance,
Key: ReversibleKeyGenerator,
Value: FullCodec,
QueryKind: QueryKindTrait<Value, OnEmpty>,
OnEmpty: 'static + Get<<QueryKind as QueryKindTrait<Value, OnEmpty>>::Query>,
MaxValues: Get<Option<u32>>,

pub fn iter_prefix<KP>(
kp: KP
) -> PrefixIterator<(<Key as HasKeyPrefix<KP>>::Suffix, Value), ()> where
Key: HasReversibleKeyPrefix<KP>,

Enumerate all elements in the map with prefix key kp in no particular order.

If you add or remove values whose prefix key is kp to the map while doing this, you’ll get undefined results.

pub fn iter_prefix_from<KP>(
kp: KP,
starting_raw_key: Vec<u8, Global>
) -> PrefixIterator<(<Key as HasKeyPrefix<KP>>::Suffix, Value), ()> where
Key: HasReversibleKeyPrefix<KP>,

Enumerate all elements in the map with prefix key kp after a specified starting_raw_key in no particular order.

If you add or remove values whose prefix key is kp to the map while doing this, you’ll get undefined results.

pub fn iter_key_prefix<KP>(
kp: KP
) -> KeyPrefixIterator<<Key as HasKeyPrefix<KP>>::Suffix> where
Key: HasReversibleKeyPrefix<KP>,

Enumerate all suffix keys in the map with prefix key kp in no particular order.

If you add or remove values whose prefix key is kp to the map while doing this, you’ll get undefined results.

pub fn iter_key_prefix_from<KP>(
kp: KP,
starting_raw_key: Vec<u8, Global>
) -> KeyPrefixIterator<<Key as HasKeyPrefix<KP>>::Suffix> where
Key: HasReversibleKeyPrefix<KP>,

Enumerate all suffix keys in the map with prefix key kp after a specified starting_raw_key in no particular order.

If you add or remove values whose prefix key is kp to the map while doing this, you’ll get undefined results.

pub fn drain_prefix<KP>(
kp: KP
) -> PrefixIterator<(<Key as HasKeyPrefix<KP>>::Suffix, Value), ()> where
Key: HasReversibleKeyPrefix<KP>,

Remove all elements from the map with prefix key kp and iterate through them in no particular order.

If you add elements with prefix key k1 to the map while doing this, you’ll get undefined results.

pub fn iter() -> PrefixIterator<(<Key as KeyGenerator>::Key, Value), ()>

Enumerate all elements in the map in no particular order.

If you add or remove values to the map while doing this, you’ll get undefined results.

pub fn iter_from(
starting_raw_key: Vec<u8, Global>
) -> PrefixIterator<(<Key as KeyGenerator>::Key, Value), ()>

Enumerate all elements in the map after a specified starting_key in no particular order.

If you add or remove values to the map while doing this, you’ll get undefined results.

pub fn iter_keys() -> KeyPrefixIterator<<Key as KeyGenerator>::Key>

Enumerate all keys in the map in no particular order.

If you add or remove values to the map while doing this, you’ll get undefined results.

pub fn iter_keys_from(
starting_raw_key: Vec<u8, Global>
) -> KeyPrefixIterator<<Key as KeyGenerator>::Key>

Enumerate all keys in the map after a specified starting_raw_key in no particular order.

If you add or remove values to the map while doing this, you’ll get undefined results.

pub fn drain() -> PrefixIterator<(<Key as KeyGenerator>::Key, Value), ()>

Remove all elements from the map and iterate through them in no particular order.

If you add elements to the map while doing this, you’ll get undefined results.

pub fn translate<O, F>(f: F)where
O: Decode,
F: FnMut(<Key as KeyGenerator>::Key, O) -> Option<Value>,

Translate the values of all elements by a function f, in the map in no particular order.

By returning None from f for an element, you’ll remove it from the map.

NOTE: If a value fail to decode because storage is corrupted then it is skipped.